chore(common): Move functions to self contained files
