+++ src/update-checkout-message.js
@@ -1,0 +1,17 @@
+//TODO: Refactor to use async await.
+import { getCookie } from './get-cookie';
+
+function updateCheckoutMessage(cartId, message) {
+    console.log(`update checkout <${cartId}> message to <${message}>`);
+
+    fetch(`/api/storefront/checkouts/${cartId}`, {
+        "headers": {
+            'Content-Type': 'application/json',
+            'x-xsrf-token': getCookie('XSRF-TOKEN'),
+        },
+        "method": "PUT",
+        "body": `${message}`
+    })
+    .then(response => response.json())
+    .then(json => console.log(json))
+}
